matverse-copilot/requirements.txt
web3>=6.0.0
watchdog>=3.0.0
tweepy>=4.14.0
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
click>=8.1.0
colorama>=0.4.6
psutil>=5.9.0
schedule>=1.2.0
py-solc-x>=2.0.0
